Restructured \pages
