+++ src/core/header/menuItems.ts
@@ -5,22 +5,24 @@ interface menuItemInterface {
 }
 
 export interface menuItemsInterface extends Array<menuItemInterface> {}
-// need to interface with /folder/pageA linking to /pageB
+//TODO: need to interface with /folder/pageA linking to /pageB
 const menuItems: menuItemInterface[] = [
   {
     text: "Home",
     uri: `/`,
   },
 
-  // need to add a "remove whitespace" so the gridtemplate columns can take e.g. `tic tac toe` as input
+  //TODO: need to add a "remove whitespace" so the gridtemplate columns can take e.g. `tic tac toe` as input
   {
     text: "TicTacToe",
-    // uri: "Games/TicTacToe",
-    uri: "/TicTacToe",
+    uri: "/Games/TicTacToe",
   },
-  { text: "connect4", uri: "/connect4" },
-  { text: "Clocks", uri: "/Clocks" },
-  { text: "Colours", uri: "/ColourSwatch" },
+  { text: "connect4", uri: "/Games/connect4" },
+
+  { text: "Clocks", uri: "/Visuals/Clocks" },
+  { text: "Colours", uri: "/Visuals/ColourSwatch" },
+
+  { text: "About", uri: "/Info/About" },
 ];
 
 export default menuItems;
