+++ src/pages/Info/About.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import BlogClass from "../../components/blogClass/BlogClass";
+
+const About: React.FC = () => {
+  return <BlogClass blogName="personal/HomePage" />;
+};
+export default About;
