+++ src/pages/Visuals/Clocks.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 
-import RoundClock from "../components/clocks/RoundClock";
-import MiniClock from "../components/clocks/MiniClock";
-import BlogClass from "../components/blogClass/BlogClass";
-import FullPageBody from "../PageTypes/FullPageBody";
-import { useTime } from "../utils/other/useTime";
+import RoundClock from "../../components/clocks/RoundClock";
+import MiniClock from "../../components/clocks/MiniClock";
+import BlogClass from "../../components/blogClass/BlogClass";
+import FullPageBody from "../../PageTypes/FullPageBody";
+import { useTime } from "../../utils/other/useTime";
 
 const NextFCClocks: React.FC<FCClocksProps> = ({}) => {
   return <FullPageBody content={<FCClocks />} />;
